splitting bulk_download controller, making UX wizard-like
