+++ app/javascript/lib/scp-api.js
@@ -79,7 +79,7 @@ function defaultPostInit(mock=false) {
 export async function fetchAuthCode(mock=false) {
   const init = defaultPostInit(mock)
 
-  const [authCode, perfTimes] = await scpApi('/search/auth_code', init, mock)
+  const [authCode, perfTimes] = await scpApi('/bulk_download/auth_code', init, mock)
 
   logDownloadAuthorization(perfTimes)
 
@@ -449,24 +449,16 @@ export async function fetchFacetFilters(facet, query, mock=false) {
 }
 
 /**
- *  Returns number of files and bytes (by file type), to preview bulk download
+ *  Queries the bulk_download/summary API to retrieve a list of study and file information
  *
  * Docs:
- * https://singlecell.broadinstitute.org/single_cell/api/swagger_docs/v1#!/Search/search_bulk_download_size_path
+ * https://singlecell.broadinstitute.org/single_cell/api/swagger_docs/v1#!/BulkDownload/bulk_download_summary_path
  *
  * @param {Array} accessions List of study accessions to preview download
- * @param {Array} fileTypes List of file types in studies to preview download
- *
- * @example returns Promise for JSON
- * {
- *  "Expression": {"total_files": 4, "total_bytes": 1797720765},
- *  "Metadata": {"total_files": 2, "total_bytes": 865371}
- * }
- * fetchDownloadSize([SCP200, SCP201], ["Expression", "Metadata"])
  */
 export async function fetchDownloadInfo(accessions, mock=false) {
   const queryString = `?accessions=${accessions}`
-  const pathAndQueryString = `/search/bulk_download_info/${queryString}`
+  const pathAndQueryString = `/bulk_download/summary/${queryString}`
   const [info] = await scpApi(pathAndQueryString, defaultInit(), mock)
   return info
 }
