+++ app/javascript/styles/_search.scss
@@ -337,6 +337,7 @@ nav.search-links, #search-panel {
 }
 
 .download-table-container {
-	height: 70vh;
 	margin: 2em 0em 1em 0em;
+	overflow-y: scroll;
+	overflow-x: hidden;
 }
